Include versions with NPM packages
The manifest does not include versions in the package names, but then when checking the manifest against the state of the application, version names are included.  This results in a bunch of "X is in the application but not the manifest" (and the opposite) errors.

This change results in the version being stored the manifest now.